Run tests only on supported Oses.
